requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
bleach==6.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.0.10
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-classy-tags==2.0.0
django-cms==3.7.4
django-debug-toolbar==3.2.2
django-environ==0.11.2
django-filer==2.0.2
django-formtools==2.2
django-htmlmin==0.11.0
django-hvad==1.8.0
django-instagram==0.3.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-nocaptcha-recaptcha==0.0.20
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-social-share==2.0.0
django-social-widgets==0.4.0
django-summernote==0.8.20.0
django-sortedm2m==3.1.1
django-summernote==0.8.20.0
django-taggit==4.0.0
django-taggit-autosuggest==0.4.1
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
docopt==0.4.0
easy-thumbnails==2.7
gunicorn==20.0.4
html5lib==1.1
idna==3.2
lxml==4.8.0
mailchimp==2.0.10
mysql-client==0.0.1
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
requests==2.26.0
six==1.15.0
sorl-thumbnail==12.8.0
soupsieve==2.2.1
South==1.0.2
sqlparse==0.3.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.26.6
webencodings==0.5.1
